requirements.txt
streamlit==1.50.0
langchain-core==0.3.79
langchain-ollama==0.3.10
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langgraph-checkpoint-sqlite==2.0.11
ollama==0.6.0
python-dotenv==1.1.1
requests==2.32.5

# Optional / common dependencies present in the venv (kept for compatibility)
httpx==0.28.1
orjson==3.11.3
pydantic==2.12.0
typing_extensions==4.15.0